[issue #20] adjusted GUI
I adjusted the GUI by:
putting the labels in JScrollPane's to prevent annoying window resizing
I added unicode escapes in the buttons
I made the sizing of calculatorButton, the JLabels, and the gaps between
JPanels more consistent.
I removed the ability to resize the calculator as a user

I also fixed a couple bugs in the SimpleCalcState class and added a few
more thrown exceptions.